+++ src/App.js
@@ -1,19 +1,19 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
-import './App.css';
+import './assets/stylesheets/App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import Login from './pages/Login';
-import Home from './pages/Home';
-import RegisterScreen from './pages/Register';
+import Login from './components/auth/Login';
+import Home from './components/home/Home';
+import RegisterScreen from './components/auth/Register';
 import Navbar from './components/nav/Navbar';
 import TopNav from './components/nav/Topbar';
-import CarDetails from './pages/carDetails/CarDetails';
-import DeleteScreen from './pages/delete/DeleteScreen';
-import ReservationList from './components/ReservationList';
-import AddReservation from './pages/reservations/AddReservation';
+import CarDetails from './components/car/CarDetails';
+import DeleteScreen from './components/car/DeleteScreen';
+import ReservationList from './components/reservation/ReservationList';
+import AddReservation from './components/reservation/AddReservation';
 
 function App() {
   const { isAuthenticated } = useSelector((state) => state.auth);
