+++ src/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import store from './app/store';
+import store from './redux/store';
 import App from './App';
 
 test('renders learn react link', () => {
