+++ src/assets/stylesheets/AddReservation.css
@@ -1,5 +1,5 @@
 .add-reservation-page {
-  background-image: url('./audi-audi-rs5-coupe-car-wallpaper-preview.jpeg');
+  background-image: url('../images//audi-audi-rs5-coupe-car-wallpaper-preview.jpeg');
   background-repeat: no-repeat;
   background-size: cover;
   width: 100%;
