+++ src/assets/stylesheets/App.css
@@ -29,7 +29,7 @@ html {
 
 .loginPage,
 .registerPage {
-  background-image: url('./assets/Best\ Car\ Wallpapers\ Hd\ 35\ Pictures.jpg');
+  background-image: url('../images/Best\ Car\ Wallpapers\ Hd\ 35\ Pictures.jpg');
   height: 100vh;
   background-position: center;
   background-repeat: no-repeat;
@@ -263,8 +263,3 @@ label a {
   height: auto;
   min-width: 400px;
 }
-
-/* .carContainer > img {
-  width: 100%;
-  height: 100%;
-} */
