+++ src/assets/stylesheets/Deletescreen.css
@@ -4,7 +4,7 @@
   /* width: 100%; */
 
   /* margin-left: 20%; */
-  background-image: url("../../assets/audi-auto-vehicle-sports-car.jpeg");
+  background-image: url("../images/audi-auto-vehicle-sports-car.jpeg");
   height: 100vh;
   background-position: center;
   background-repeat: no-repeat;
