+++ src/components/nav/Navbar.js
@@ -6,12 +6,12 @@ import {
   FaTwitter, FaFacebookF, FaPinterestP, FaVimeoV,
 } from 'react-icons/fa';
 import { TiSocialGooglePlus, TiThMenu, TiTimes } from 'react-icons/ti';
-import './nav.css';
+import '../../assets/stylesheets/nav.css';
 
 const adminLinks = [
   { id: 1, path: '/', text: 'MODELS' },
   { id: 2, path: '/addReservation', text: 'RESERVE' },
-  { id: 3, path: '/reservation', text: 'RESERVATIONS' },
+  { id: 3, path: '/reservationList', text: 'RESERVATIONS' },
   { id: 4, path: '/add-car', text: 'ADD CAR' },
   { id: 5, path: '/delete', text: 'DELETE' },
 ];
@@ -19,7 +19,7 @@ const adminLinks = [
 const links = [
   { id: 1, path: '/', text: 'MODELS' },
   { id: 2, path: '/addReservation', text: 'RESERVE' },
-  { id: 3, path: '/reservation', text: 'MY RESERVATION' },
+  { id: 3, path: '/reservationList', text: 'MY RESERVATION' },
 ];
 
 const SocialMedia = () => (
